+++ packages/core/src/hooks/useBlockMeta.ts
@@ -4,6 +4,7 @@ import { useMulticallAddress } from './useMulticallAddress'
 import { QueryParams } from '../constants/type/QueryParams'
 import { useRawCall } from './useRawCalls'
 import { useChainId } from './useChainId'
+import { useConfig } from './useConfig'
 
 const GET_CURRENT_BLOCK_TIMESTAMP_CALL = MultiCallABI.encodeFunctionData('getCurrentBlockTimestamp', [])
 const GET_CURRENT_BLOCK_DIFFICULTY_CALL = MultiCallABI.encodeFunctionData('getCurrentBlockDifficulty', [])
@@ -14,13 +15,31 @@ const GET_CURRENT_BLOCK_DIFFICULTY_CALL = MultiCallABI.encodeFunctionData('getCu
  */
 export function useBlockMeta(queryParams: QueryParams = {}) {
   const chainId = useChainId({ queryParams })
+  const { refresh: configRefresh } = useConfig()
 
   const address = useMulticallAddress(queryParams)
+  const refresh = queryParams.refresh ?? configRefresh
+  const isStatic = queryParams.isStatic ?? refresh === 'never'
+  const refreshPerBlocks = typeof refresh === 'number' ? refresh : undefined
   const timestamp = useRawCall(
-    address && chainId !== undefined && { address, data: GET_CURRENT_BLOCK_TIMESTAMP_CALL, chainId }
+    address &&
+      chainId !== undefined && {
+        address,
+        data: GET_CURRENT_BLOCK_TIMESTAMP_CALL,
+        chainId,
+        isStatic,
+        refreshPerBlocks,
+      }
   )
   const difficulty = useRawCall(
-    address && chainId !== undefined && { address, data: GET_CURRENT_BLOCK_DIFFICULTY_CALL, chainId }
+    address &&
+      chainId !== undefined && {
+        address,
+        data: GET_CURRENT_BLOCK_DIFFICULTY_CALL,
+        chainId,
+        isStatic,
+        refreshPerBlocks,
+      }
   )
 
   return {