+++ packages/core/src/hooks/useConfig.test.tsx
@@ -31,7 +31,7 @@ describe('useConfig', () => {
     const { result, waitForCurrent } = await renderDAppHook(() => useConfig(), { config: setup.config })
     await waitForCurrent((val) => val !== undefined)
     expect(result.error).to.be.undefined
-    expect(result.current.networks?.length).to.eq(42)
+    expect(result.current.networks?.length).to.eq(43)
     expect(result.current.notifications?.checkInterval).to.eq(500)
     expect(result.current.notifications?.expirationPeriod).to.eq(5000)
   })