+++ packages/core/src/providers/network/readonlyNetworks/context.ts
@@ -1,8 +1,15 @@
 import { createContext, useContext } from 'react'
-import { Providers } from './model'
+import { ReadonlyNetworksModel } from './model'
 
-export const ReadonlyNetworksContext = createContext<Providers>({})
+export const ReadonlyNetworksContext = createContext<ReadonlyNetworksModel>({
+  providers: {},
+  updateNetworkState: () => undefined,
+})
 
 export function useReadonlyNetworks() {
-  return useContext(ReadonlyNetworksContext)
+  return useContext(ReadonlyNetworksContext).providers
+}
+
+export function useUpdateNetworksState() {
+  return useContext(ReadonlyNetworksContext).updateNetworkState
 }