+++ packages/core/src/testing/renderWeb3Hook.tsx
@@ -1,13 +1,14 @@
 import { MockProvider } from '@ethereum-waffle/provider'
 import { renderHook } from '@testing-library/react-hooks'
-import { BlockNumberProvider, NetworkProvider, MultiChainStateProvider } from '../providers'
+import { BlockNumberProvider, NetworkProvider, MultiChainStateProvider, ConfigProvider } from '../providers'
 import React from 'react'
 import { deployMulticall, getWaitUtils, IdentityWrapper, mineBlock } from './utils'
 import { BlockNumbersProvider } from '../providers/blockNumber/blockNumbers'
 import { ReadonlyNetworksProvider } from '../providers/network'
 
 export interface renderWeb3HookOptions<Tprops> {
-  mockProvider?: MockProvider | Record<number /* ChainId */, MockProvider>
+  mockProvider?: MockProvider
+  readonlyMockProviders?: Record<number /* ChainId */, MockProvider>
   mockProviderOptions?: {
     pollingInterval?: number
   }
@@ -38,11 +39,9 @@ export const renderWeb3Hook = async <Tprops, TResult>(
 ) => {
   const providers: Record<number, MockProvider> = {}
   const multicallAddresses: Record<number, string> = {}
-  let defaultProvider = new MockProvider()
 
   const addSingleProvider = async (currentProvider: MockProvider) => {
     const { chainId } = await currentProvider.getNetwork()
-    currentProvider.pollingInterval = options?.mockProviderOptions?.pollingInterval ?? 200
     providers[chainId] = currentProvider
 
     const mockMulticallAddresses = await deployMulticall(currentProvider, chainId)
@@ -54,32 +53,39 @@ export const renderWeb3Hook = async <Tprops, TResult>(
     await currentProvider.getBlockNumber()
   }
 
-  const providerObject = options?.mockProvider || new MockProvider()
-  if (providerObject instanceof MockProvider) {
-    defaultProvider = providerObject
-    await addSingleProvider(providerObject)
-  } else {
-    for (const chainIdString in providerObject) {
-      const chainId = Number(chainIdString)
-      await addSingleProvider(providerObject[chainId])
-    }
+  const defaultProvider = options?.mockProvider || new MockProvider()
+  await addSingleProvider(defaultProvider)
+
+  const readOnlyProviders = options?.readonlyMockProviders ?? {}
+  for (const chainIdString in readOnlyProviders) {
+    const chainId = Number(chainIdString)
+    await addSingleProvider(readOnlyProviders[chainId])
+  }
+
+  if (Object.keys(readOnlyProviders).length === 0) {
+    const defaultReadOnlyProvider = new MockProvider()
+    await addSingleProvider(defaultReadOnlyProvider)
+    const { chainId } = await defaultReadOnlyProvider.getNetwork()
+    readOnlyProviders[chainId] = defaultReadOnlyProvider
   }
 
   const UserWrapper = options?.renderHook?.wrapper ?? IdentityWrapper
 
   const { result, waitForNextUpdate, rerender, unmount } = renderHook<Tprops, TResult>(hook, {
     wrapper: (wrapperProps) => (
-      <NetworkProvider providerOverride={defaultProvider}>
-        <ReadonlyNetworksProvider providerOverrides={providers}>
-          <BlockNumberProvider>
-            <BlockNumbersProvider>
-              <MultiChainStateProvider multicallAddresses={multicallAddresses}>
-                <UserWrapper {...wrapperProps} />
-              </MultiChainStateProvider>
-            </BlockNumbersProvider>
-          </BlockNumberProvider>
-        </ReadonlyNetworksProvider>
-      </NetworkProvider>
+      <ConfigProvider config={{ pollingInterval: options?.mockProviderOptions?.pollingInterval ?? 200 }}>
+        <NetworkProvider providerOverride={defaultProvider}>
+          <ReadonlyNetworksProvider providerOverrides={readOnlyProviders}>
+            <BlockNumberProvider>
+              <BlockNumbersProvider>
+                <MultiChainStateProvider multicallAddresses={multicallAddresses}>
+                  <UserWrapper {...wrapperProps} />
+                </MultiChainStateProvider>
+              </BlockNumbersProvider>
+            </BlockNumberProvider>
+          </ReadonlyNetworksProvider>
+        </NetworkProvider>
+      </ConfigProvider>
     ),
     initialProps: options?.renderHook?.initialProps,
   })
@@ -87,7 +93,9 @@ export const renderWeb3Hook = async <Tprops, TResult>(
   return {
     result,
     defaultProvider,
-    mineBlock: async () => mineBlock(defaultProvider),
+    mineBlock: async () => {
+      await Promise.all([defaultProvider, ...Object.values(readOnlyProviders)].map((provider) => mineBlock(provider)))
+    },
     rerender,
     unmount,
     // do not return the waitFor* functions from `renderHook` - they are not usable after using waitForNextUpdate().